# src/main/resources/openapi/openapi-garden.yaml
openapi: 3.0.3
info:
  title: Bio-Garden API
  version: 1.0.0
  description: OpenAPI definition to backend app in Spring Boot.
servers:
  - url: https://localhost:8443
    description: Local server (HTTPS)
  - url: http://localhost:8080
    description: Local server (HTTP – redirects to HTTPS)

tags:
  - name: Plants
    description: Operations related to plants
  - name: PlantCompanions
    description: Operations related to plant companion relationships

paths:
  /api/v1/plants:
    get:
      tags:
        - Plants
      summary: Download paginated list of plants
      description: Returns paginated list of plants.
      operationId: listPlants
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Page size (number of items per page)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - name: sort
          in: query
          description: Sorting \. `name,asc`
          required: false
          schema:
            type: string
            default: name,asc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantPage"

        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags:
        - Plants
      summary: Create new plant
      operationId: createPlant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantCreateRequest"

      responses:
        "201":
          description: Created
          headers:
            Location:
              description: URL to newly created plant
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/plants/with-companions:
    get:
      tags:
        - Plants
      summary: Download paginated list of plants with companion relationships
      operationId: listPlantsWithCompanions
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: name,asc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantWithCompanionsPage"
        "400":
          $ref: "#/components/responses/BadRequest"

  /api/v1/plants/{plantId}:
    get:
      tags:
        - Plants
      summary: Download plant by ID
      operationId: getPlantById
      parameters:
        - name: plantId
          in: path
          required: true
          description: plant ID
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Plants
      summary: Update plant
      operationId: updatePlant
      parameters:
        - name: plantId
          in: path
          required: true
          description: plant ID
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags:
        - Plants
      summary: Remove plant
      operationId: deletePlant
      parameters:
        - name: plantId
          in: path
          required: true
          description: Plant ID
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/plant-companions:
    post:
      tags:
        - PlantCompanions
      summary: Create companion relationship
      operationId: createCompanionRelationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantCompanion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/plant-companions/{id}:
    put:
      tags:
        - PlantCompanions
      summary: Update companion relationship
      operationId: updateCompanionRelationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanionUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantCompanion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - PlantCompanions
      summary: Delete companion relationship
      operationId: deleteCompanionRelationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/plant-companions/plant/{plantId}:
    get:
      tags:
        - PlantCompanions
      summary: Get all companions for a plant
      operationId: getCompanionsForPlant
      parameters:
        - name: plantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantCompanion"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/plant-companions/plant/{plantId}/type/{relationshipType}:
    get:
      tags:
        - PlantCompanions
      summary: Get companions by relationship type
      operationId: getCompanionsByType
      parameters:
        - name: plantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: relationshipType
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/RelationshipType"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantCompanion"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

    Conflict:
      description: Conflict (e.g. duplicate name)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  schemas:
    Plant:
      type: object
      required:
        - id
        - name
        - crop
        - sowFrom
        - sowTo
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          minLength: 1
          maxLength: 120
          example: "Cucumber"
        crop:
          $ref: "#/components/schemas/CropType"
        sowPeriod:
          $ref: "#/components/schemas/SowPeriod"
        daysToHarvest:
          type: integer
          minimum: 1
          example: 60
        creationDate:
          type: string
          format: date-time
          example: "2026-01-28T10:15:30Z"
        lastUpdateDate:
          type: string
          format: date-time
          example: "2026-02-28T10:15:30Z"
    CropType:
      type: string
      enum:
        - LEAF
        - ROOT
        - FRUIT
        - FLOWER
        - SEED
    SowPeriod:
      type: object
      properties:
        sowFrom:
          type: string
          format: date-time
        sowTo:
          type: string
          format: date-time
    PlantCreateRequest:
      type: object
      required:
        - name
        - crop
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        crop:
           $ref: "#/components/schemas/CropType"
        sowPeriod:
          $ref: "#/components/schemas/SowPeriod"
        daysToHarvest:
          type: integer
          minimum: 1

    PlantUpdateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        crop:
          $ref: "#/components/schemas/CropType"
        sowPeriod:
          $ref: "#/components/schemas/SowPeriod"
        daysToHarvest:
          type: integer
          minimum: 1
    RelationshipType:
      type: string
      enum:
        - GOOD
        - BAD
        - NEUTRAL
        - COMPANION_ROW

    PlantCompanion:
      type: object
      required:
        - id
        - plantId
        - companionPlantId
        - relationshipType
      properties:
        id:
          type: integer
          format: int64
          example: 1
        plantId:
          type: integer
          format: int64
          example: 1
        companionPlantId:
          type: integer
          format: int64
          example: 2
        relationshipType:
          $ref: "#/components/schemas/RelationshipType"
        recommendedDistanceCm:
          type: integer
          minimum: 1
          example: 30
        bidirectional:
          type: boolean
          example: true

    CompanionRequest:
      type: object
      required:
        - plantId
        - companionPlantId
        - relationshipType
      properties:
        plantId:
          type: integer
          format: int64
        companionPlantId:
          type: integer
          format: int64
        relationshipType:
          $ref: "#/components/schemas/RelationshipType"
        recommendedDistanceCm:
          type: integer
          minimum: 1
        bidirectional:
          type: boolean

    CompanionUpdateRequest:
      type: object
      properties:
        recommendedDistanceCm:
          type: integer
          minimum: 1
        bidirectional:
          type: boolean
        relationshipType:
          $ref: "#/components/schemas/RelationshipType"

    PlantPage:
      type: object
      required:
        - content
        - page
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Plant"
        page:
          $ref: "#/components/schemas/PageInfo"

    PageInfo:
      type: object
      required:
        - number
        - size
        - totalElements
        - totalPages
      properties:
        number:
          type: integer
          minimum: 0
          example: 0
        size:
          type: integer
          minimum: 1
          example: 20
        totalElements:
          type: integer
          minimum: 0
          example: 1
        totalPages:
          type: integer
          minimum: 0
          example: 1

    PlantWithCompanions:
      allOf:
        - $ref: "#/components/schemas/Plant"
        - type: object
          properties:
            companions:
              type: array
              items:
                $ref: "#/components/schemas/PlantCompanion"

    PlantWithCompanionsPage:
      type: object
      required:
        - content
        - page
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PlantWithCompanions"
        page:
          $ref: "#/components/schemas/PageInfo"

    Problem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: "https://example.com/problems/validation"
        title:
          type: string
          example: "Validation failed"
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: "Feld `name` is required"
        instance:
          type: string
          example: "/api/v1/plants"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"

    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: "name"
        message:
          type: string
          example: "must not be blank"